move debug print func to dev app
